backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
requests==2.25.0
